07/16 방 나가기 코드 수정

+++ Source/Client/DBConnector.py
@@ -185,29 +185,32 @@ class DBConnector:      # DB를 총괄하는 클래스
 
         return _cr_id
 
-    def delete_my_table(self, data:DeleteMyTable):
+    def delete_my_table(self, data:DeleteTable):
         """PK,FK를 고려하여 순서작성함"""
 
-        #CTB_USER_CHATROOM의 CR_ID에 해당하는 내용 삭제
-        sql_1 = f"DELETE FROM CTB_USER_CHATROOM WHERE CR_ID = {data.cr_id}"
-        #CTB_CHATROOM 의 CR_ID 삭제
-        sql_2 = f"DELETE FROM CTB_CHATROOM WHERE CR_ID = {data.cr_id}"
-        #CTB_READ_CNT_{CR_ID} 테이블 삭제
-        sql_3 = f"DROP TABLE CTB_READ_CNT_{data.cr_id}"
-        #CTB_CONTENT_{CR_ID} 테이블 삭제
-        sql_4 = f"DROP TABLE CTB_CONTENT_{data.cr_id}"
-
-        process_sql = [sql_1, sql_2, sql_3, sql_4]
-        try:
-            for sql in process_sql:
-                self.conn.execute(sql)
+        if self.user_id != data.my_id:
+            # CTB_CHATROOM 의 CR_ID 삭제
+            sql_2 = f"DELETE FROM CTB_CHATROOM WHERE CR_ID = '{data.cr_id}' and USER_ID = '{data.my_id}'"
+            self.conn.execute(sql_2)
             self.commit_db()
-        except Exception as e:
-            self.conn.rollback()
-            # 오류 처리 또는 로깅
-            print(f"delete_my_table에서 Error occurred: {e}")
-        finally:
-            self.end_conn()
+        else:
+            sql_2 = f"DELETE FROM CTB_CHATROOM WHERE CR_ID = '{data.cr_id}'"
+            #CTB_USER_CHATROOM의 CR_ID에 해당하는 내용 삭제
+            sql_1 = f"DELETE FROM CTB_USER_CHATROOM WHERE CR_ID = '{data.cr_id}'"
+            #CTB_READ_CNT_{CR_ID} 테이블 삭제
+            sql_3 = f"DROP TABLE CTB_READ_CNT_{data.cr_id}"
+            #CTB_CONTENT_{CR_ID} 테이블 삭제
+            sql_4 = f"DROP TABLE CTB_CONTENT_{data.cr_id}"
+
+            process_sql = [sql_1, sql_2, sql_3, sql_4]
+            try:
+                for sql in process_sql:
+                    self.conn.execute(sql)
+                self.commit_db()
+            except Exception as e:
+                self.conn.rollback()
+                # 오류 처리 또는 로깅
+                print(f"delete_my_table에서 Error occurred: {e}")
 
     ## TB_content ================================================================================ ##
     # 대화 추가
